+++ rust/src/24/24480.rs
@@ -11,8 +11,8 @@ fn main() {
     let mut adjacency_list = vec![Vec::new(); n + 1];
 
     for (u, v) in (0..m).map(|_| (input(), input())) {
-        adjacency_list[u].push(v);
-        adjacency_list[v].push(u);
+        adjacency_list[u].push(v as i32);
+        adjacency_list[v].push(u as i32);
     }
 
     for list in adjacency_list.iter_mut() {
@@ -28,20 +28,20 @@ fn main() {
     print!("{output}");
 }
 
-fn dfs(graph: &[Vec<usize>], start: usize) -> Vec<usize> {
+fn dfs(graph: &[Vec<i32>], start: usize) -> Vec<i32> {
     let mut visited = vec![0; graph.len()];
-    let mut stack = vec![start];
+    let mut stack = vec![start as i32];
     let mut count = 1;
 
     while let Some(node) = stack.pop() {
-        if visited[node] != 0 {
+        if visited[node as usize] != 0 {
             continue;
         }
 
-        visited[node] = count;
+        visited[node as usize] = count;
         count += 1;
 
-        for &neighbor in graph[node].iter() {
+        for &neighbor in graph[node as usize].iter() {
             stack.push(neighbor);
         }
     }
